pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "devtools>=0.12.2",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "logfire>=3.5.3",
    "pydantic-ai>=0.0.23",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
